Backend/requirements.txt
flask 
pyserial